pyproject.toml
[project]
name = "conversational-finance-ultra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "accelerate==1.11.0",
    "acres==0.5.0",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anthropic==0.72.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "authlib==1.6.5",
    "backoff==2.2.1",
    "beartype==0.22.4",
    "brotli==1.1.0",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "chonkie==1.2.1",
    "ci-info==0.3.0",
    "click==8.3.0",
    "cloudpickle==3.1.1",
    "configobj==5.0.9",
    "configparser==7.2.0",
    "cryptography==46.0.3",
    "cyclopts==4.0.0",
    "datasets==4.3.0",
    "deprecation==2.1.0",
    "dill==0.4.0",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "docutils==0.22.2",
    "easyocr==1.7.2",
    "ecdsa==0.19.1",
    "einops==0.8.1",
    "email-validator==2.3.0",
    "etelemetry==0.3.1",
    "fastapi==0.115.14",
    "fastapi-cli==0.0.14",
    "fastapi-cloud-cli==0.3.1",
    "fastmcp==2.13.0.1",
    "ffmpy==0.6.4",
    "filelock==3.20.0",
    "frozenlist==1.8.0",
    "fsspec==2025.9.0",
    "groovy==0.1.2",
    "grpcio==1.76.0",
    "h11==0.16.0",
    "h2==4.3.0",
    "hf-xet==1.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httplib2==0.31.0",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.36.0",
    "hyperframe==6.1.0",
    "idna==3.11",
    "ijson==3.4.0.post0",
    "imageio==2.37.0",
    "ipywidgets==8.1.7",
    "isodate==0.7.2",
    "itsdangerous==2.2.0",
    "jaraco-classes==3.4.0",
    "jaraco-context==6.0.1",
    "jaraco-functools==4.3.0",
    "jinja2==3.1.6",
    "jiter==0.11.1",
    "joblib==1.5.2",
    "json-repair==0.52.3",
    "jsonschema==4.25.1",
    "jsonschema-path==0.3.4",
    "jsonschema-specifications==2025.9.1",
    "jupyterlab-widgets==3.0.15",
    "keyring==25.6.0",
    "lazy-loader==0.4",
    "lazy-object-proxy==1.12.0",
    "loguru==0.7.3",
    "looseversion==1.3.0",
    "lxml==6.0.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mcp==1.19.0",
    "mdurl==0.1.2",
    "mem0ai==1.0.0",
    "memoryos==1.1.2",
    "mlx==0.29.3",
    "mlx-lm==0.28.3",
    "mlx-metal==0.29.3",
    "mlx-vlm==0.3.5",
    "more-itertools==10.8.0",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "multiprocess==0.70.16",
    "neo4j==6.0.2",
    "networkx==3.5",
    "nibabel==5.3.2",
    "ninja==1.13.0",
    "nipype==1.10.0",
    "numpy==2.2.6",
    "ollama==0.4.9",
    "openai==1.109.1",
    "openapi-core==0.19.5",
    "openapi-pydantic==0.5.1",
    "openapi-schema-validator==0.6.3",
    "openapi-spec-validator==0.7.2",
    "opencv-python==4.12.0.88",
    "opencv-python-headless==4.12.0.88",
    "orjson==3.11.4",
    "paddleocr>=2.0.6",
    "pandas==2.3.3",
    "parse==1.20.2",
    "pathable==0.4.4",
    "pathlib==1.0.1",
    "pathvalidate==3.3.1",
    "pdf2image==1.17.0",
    "pika==1.3.2",
    "pillow==11.3.0",
    "portalocker==3.2.0",
    "postgrest==2.23.0",
    "posthog==6.7.11",
    "propcache==0.4.1",
    "prov==2.1.1",
    "puremagic==1.30",
    "py-key-value-aio==0.2.8",
    "py-key-value-shared==0.2.8",
    "pyarrow==22.0.0",
    "pyasn1==0.6.1",
    "pyclipper==1.3.0.post6",
    "pycparser==2.23",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-extra-types==2.10.6",
    "pydantic-settings==2.11.0",
    "pydot==4.0.1",
    "pydub==0.25.1",
    "pyjwt==2.10.1",
    "pymupdf==1.26.5",
    "pymysql==1.1.2",
    "pyparsing==3.2.5",
    "pypdf==6.1.3",
    "pyperclip==1.11.0",
    "pytesseract==0.3.13",
    "python-bidi==0.6.7",
    "python-dotenv==1.2.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyxnat==1.6.3",
    "pyyaml==6.0.3",
    "qdrant-client==1.15.1",
    "rdflib==7.4.0",
    "realtime==2.23.0",
    "redis==7.0.1",
    "referencing==0.36.2",
    "regex==2025.10.23",
    "requests==2.32.5",
    "rfc3339-validator==0.1.4",
    "rich==14.2.0",
    "rich-rst==1.3.2",
    "rich-toolkit==0.15.1",
    "rignore==0.7.1",
    "rpds-py==0.28.0",
    "rsa==4.9.1",
    "ruff==0.14.2",
    "safehttpx==0.1.7",
    "safetensors==0.6.2",
    "schedule==1.2.2",
    "scikit-image==0.25.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "semantic-version==2.10.0",
    "sentence-transformers==5.1.2",
    "sentry-sdk==2.42.1",
    "setuptools==80.9.0",
    "shapely==2.1.2",
    "shellingham==1.5.4",
    "simplejson==3.20.2",
    "sniffio==1.3.1",
    "soundfile==0.13.1",
    "sqlalchemy==2.0.44",
    "sse-starlette==3.0.2",
    "starlette==0.46.2",
    "storage3==2.23.0",
    "strenum==0.4.15",
    "stripe==13.1.1",
    "supabase==2.23.0",
    "supabase-auth==2.23.0",
    "supabase-functions==2.23.0",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tifffile==2025.10.16",
    "tiktoken==0.12.0",
    "tokenizers==0.22.1",
    "tomlkit==0.13.3",
    "torch==2.9.0",
    "torchvision==0.24.0",
    "tqdm==4.67.1",
    "traits==7.0.2",
    "transformers==4.57.1",
    "typer==0.20.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "ujson==5.11.0",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==15.0.1",
    "werkzeug==3.1.1",
    "wheel==0.45.1",
    "widgetsnbextension==4.0.14",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    # fast_ocr dependencies
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "odfpy>=1.4.1",
    "striprtf>=0.0.22",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
    "chardet>=5.0.0",
    "py7zr>=0.20.0",
    "onnxruntime>=1.23.2",
    "rapidocr>=3.4.2",
]
